Set use_spark=False when cache_dir is not shared
