+++ src/datasets/packaged_modules/spark/spark.py
@@ -1,6 +1,5 @@
 import os
 import posixpath
-import uuid
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Iterable, Optional, Tuple, Union
 
@@ -14,6 +13,7 @@ from datasets.filesystems import (
     rename,
 )
 from datasets.utils.py_utils import convert_file_size_to_int
+from datasets.utils.spark_utils import validate_cache_dir
 
 
 logger = datasets.utils.logging.get_logger(__name__)
@@ -42,41 +42,13 @@ class Spark(datasets.DatasetBuilder):
 
         self._spark = pyspark.sql.SparkSession.builder.getOrCreate()
         self.df = df
-        self._validate_cache_dir(cache_dir)
+
+        validate_cache_dir(cache_dir, "Dataset.from_spark")
 
         super().__init__(
             cache_dir=cache_dir,
             config_name=str(self.df.semanticHash()),
             **config_kwargs,
-        )
-
-    def _validate_cache_dir(self, cache_dir):
-        # Returns the path of the created file.
-        def create_cache_and_write_probe(context):
-            # makedirs with exist_ok will recursively create the directory. It will not throw an error if directories
-            # already exist.
-            os.makedirs(cache_dir, exist_ok=True)
-            probe_file = os.path.join(cache_dir, "fs_test" + uuid.uuid4().hex)
-            # Opening the file in append mode will create a new file unless it already exists, in which case it will not
-            # change the file contents.
-            open(probe_file, "a")
-            return [probe_file]
-
-        if self._spark.conf.get("spark.master", "").startswith("local"):
-            return
-
-        # If the cluster is multi-node, make sure that the user provided a cache_dir and that it is on an NFS
-        # accessible to the driver.
-        # TODO: Stream batches to the driver using ArrowCollectSerializer instead of throwing an error.
-        if cache_dir:
-            probe = (
-                self._spark.sparkContext.parallelize(range(1), 1).mapPartitions(create_cache_and_write_probe).collect()
-            )
-            if os.path.isfile(probe[0]):
-                return
-
-        raise ValueError(
-            "When using Dataset.from_spark on a multi-node cluster, the driver and all workers should be able to access cache_dir"
         )
 
     def _info(self):
